Refactor: Extract badge generation logic (shield-generation)
Extracted badge generation logic into a dedicated
`utils/badges.go` file. Updated documentation to
reflect the changes and added a new font option.
Improved error handling and input validation for font
size, preventing potential panics.
