+++ docs/index.html
@@ -447,10 +447,10 @@ function clicked() {
                 <li>Simple styles: plastic-simple, flat-simple, flat-square-simple</li>
             </ul>
         </li>
-        <li><code>fontsize=11</code>: Font size for the shield text (default: 11) - must be a positive integer</li>
+        <li><code>fontsize=11</code>: Font size for the shield text (default: 11) - must be greater 2</li>
         <li><code>font=verdana</code>: Font family for the shield text (default: verdana)
             <ul>
-                <li>Available fonts: verdana, verdana-bold, verdana-bold-italic, arial, arial-bold, arial-italic, arial-bold-italic, courier-new</li>
+                <li>Available fonts: verdana, verdana-bold, verdana-bold-italic, arial, arial-bold, arial-italic, arial-bold-italic, courier-new, jetbrains-mono</li>
             </ul>
         </li>
     </ul>
@@ -499,10 +499,10 @@ function clicked() {
                 <li>Simple styles: plastic-simple, flat-simple, flat-square-simple</li>
             </ul>
         </li>
-        <li><code>fontsize=11</code>: Font size for the shield text (default: 11) - must be a positive integer</li>
+        <li><code>fontsize=11</code>: Font size for the shield text (default: 11) - must greater than 2</li>
         <li><code>font=verdana</code>: Font family for the shield text (default: verdana)
             <ul>
-                <li>Available fonts: verdana, verdana-bold, verdana-bold-italic, arial, arial-bold, arial-italic, arial-bold-italic, courier-new</li>
+                <li>Available fonts: verdana, verdana-bold, verdana-bold-italic, arial, arial-bold, arial-italic, arial-bold-italic, courier-new, jetbrains-mono</li>
             </ul>
         </li>
     </ul>
